Module-7: Moved command use cases to separate package.
